Fixed the nudge feature to auto generate context and question
Now when someone creates a nudge, the context and question are generated automatically for a better customer experience overall.

This included switching the model used for context and question generation to Gemini Flash 2.5 Lite in the backend agent service. I made this switch because we'll be making more calls by default with this feature and Gemini is half the price of GPT 4o mini.